Project 6/requirements.txt
ccxt==4.0.91
pandas==2.1.3
numpy==1.24.3
matplotlib==3.8.2
python-dotenv==1.0.0 